core/kernel batch 2: conversions, exit status, Kernel method presence
Second batch of core/kernel spec fixes (118 failures / 61 errors ->
97 / 54):

- Kernel#Float preserves identity for Float inputs and #to_f results
  (Float(nan).equal?(nan) holds, as in CRuby).
- SystemExit carries the full Integer status (the kind stored a u8, so
  exit(-2**16) reported 0 and exit(-2.2) reported 254); the process
  still exits with the OS-truncated value.
- Kernel#Array follows CRuby's nil protocol: a nil #to_ary falls
  through to #to_a, a nil #to_a wraps the argument.
- Kernel#String raises TypeError (not NoMethodError) when #to_s is
  missing or #respond_to? denies it — every probe is guarded so
  BasicObject instances reach the TypeError too.
- Kernel#initialize_copy checks the receiver's frozen state and reports
  CRuby's exact TypeError message.
- Kernel#global_variables lists the registered globals (was undefined).
- Kernel#select (IO.select), #syscall (NotImplementedError), #readline
  and #readlines (over ARGF) are defined as module functions, fixing
  the private-instance/public-singleton presence specs.
- `fail` is a true alias of `raise` on both the instance and singleton
  sides (ruby/spec compares the Method objects).
- Kernel#sleep accepts any object responding to #divmod (CRuby's
  rb_time_interval protocol).

No regressions: core/exception 0F, core/thread 1F (known), core/io
unchanged, core/file -2, full cargo test green.

+++ monoruby/src/globals.rs
@@ -821,7 +821,7 @@ impl Globals {
                     err.show_error_message_and_all_loc(&self.store);
                 }
                 Err(MonorubyErr::new(
-                    MonorubyErrKind::SystemExit(status as u8),
+                    MonorubyErrKind::SystemExit(status as i64),
                     "exit",
                 ))
             }
